Breaking tests into two files
